refactor (MyWay): Separate each section into a component
